Scale fix

+++ src/org/darkmentat/GuitarScalesBoxes/Controls/GuitarView/DisplayerFretBoard.java
@@ -6,8 +6,8 @@ interface DisplayerFretBoard
 {
     public void draw(Canvas canvas);
 
-    public int getWidth();
-    public int getHeight();
+    public float getWidth();
+    public float getHeight();
 
     public void setFretBoard(FretBoard fretBoard);
     public void setScreenSize(int width, int height);
